.kiro/specs/参考/image_backend/image_backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
google-genai>=0.7.0

google-auth>=2.30.0
requests>=2.32.0
python-multipart==0.0.9
